Add an implementation of the Dijkstra algorithm
* feat: implement Dijkstra algorithm

* refactor: add a property and replace methods

Add `statusTable` property for the  correct interaction `getAllPaths` and `getPath` methods. Add `checkAndGet_` methods for avoiding usage `!!` and replace them in `Utils.kt`.

* test: add tests for the Dijkstra algo and `Utils.kt`

* style: rename file and some variables

* docs: add doc-strings for the Dijkstra algo

Add doc-strings for `DijkstraAlgorithm` and `checkAndGet_` in `Utils.kt`.

* refactor: add the exception and the threshold in complexity

Add the exception if `idTarget` is not contained in the graph and update doc-strings and tests in the such case. Set the threshold to 18 in compexity in CyclomaticComplexMethod.

* refactor: amend the code after the review