Refactor and fix library and application projects
* fix: LeaderRank algorithm
Fix a bug where the algorithm runs endlessly if there are 0 edges in the graph. Add tests for this case.

* test: update test for LeaderRank
Update "create graph matrix of unweighted graph" test: add a check for working with edges that have negative weights.

* refactor: change the display of buttons for available algorithms
Display buttons only for those algorithms that are suitable for this type of graph.

* fix: vertex colors changing incorrectly
When switching to another vertex, the color of the previous one is restored. If a vertex is selected and the clustering algorithm is launched or the graph is reset to default, the color of the current vertex is displayed correctly.